+++ packages/docx-templates/src/mainBrowser.js
@@ -41,7 +41,7 @@ const createReport = async (options: UserOptionsInternal) => {
     processLineBreaks:
       options.processLineBreaks != null ? options.processLineBreaks : true,
     noSandbox: options.noSandbox || false,
-    vm2Sandbox: options.vm2Sandbox || false,
+    runJs: options.runJs,
     additionalJsContext: options.additionalJsContext || {},
   };
   const xmlOptions = { literalXmlDelimiter };