+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+* **Breaking change for users of the `vm2` sandbox**: replaced `vm2Sandbox` option (which caused headaches for users in the browser) with `runJs`, a custom hook for JS snippet execution. If you want to use vm2:
+
+```js
+import createReport from 'docx-templates';
+import { VM, VMScript } from 'vm2';
+
+createReport({
+  template: /* ... */,
+  data: /* ... */,
+  runJs: ({ sandbox }) => {
+    const script = new VMScript(
+      `
+      __result__ = eval(__code__);
+      `
+    ).compile();
+    const vm2 = new VM({ sandbox });
+    vm2.run(script);
+    const { VMError, Buffer, ...modifiedSandbox } = vm2._context;
+    const result = modifiedSandbox.__result__;
+    return { modifiedSandbox, result };
+  }
+})
+```
+
 ## 2.10.0 (2019-2-16)
 
 * Let alternate text be specified for images (@emilong, #57).